% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_authors.R
\name{get_authors}
\alias{get_authors}
\title{Get authors}
\usage{
get_authors()
}
\arguments{
\item{live}{logical value indicating if list of authors should be retrieved live, or from internal data.}
}
\value{
None
}
\description{
Retrieves a list of available authors in Text Collections
}
\examples{
get_authors()

}
