% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categories.R
\name{get_categories}
\alias{get_categories}
\title{Get categories}
\usage{
get_categories()
}
\arguments{
\item{live}{logical value indicating if list of authors should be retrieved live, or from internal data.}
}
\value{
None
}
\description{
FIXME
}
\examples{
get_categories()

}
