% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collections.R
\name{get_collections}
\alias{get_collections}
\title{Get collections}
\usage{
get_collections()
}
\arguments{
\item{live}{logical value indicating if list of authors should be retrieved live, or from internal data.}
}
\value{
None
}
\description{
FIXME
}
\examples{
get_collections()

}
