% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_editions.R
\name{get_editions}
\alias{get_editions}
\title{Get editions}
\usage{
get_editions()
}
\arguments{
\item{live}{logical value indicating if list of authors should be retrieved live, or from internal data.}
}
\value{
None
}
\description{
FIXME
}
\examples{
get_editions()

}
