% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genres.R
\name{get_genres}
\alias{get_genres}
\title{Get genres}
\usage{
get_genres()
}
\arguments{
\item{live}{logical value indicating if list of authors should be retrieved live, or from internal data.}
}
\value{
None
}
\description{
FIXME
}
\examples{
get_genres()

}
