% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subcollections.R
\name{get_subcollections}
\alias{get_subcollections}
\title{Get subcollections}
\usage{
get_subcollections()
}
\arguments{
\item{live}{logical value indicating if list of authors should be retrieved live, or from internal data.}
}
\value{
None
}
\description{
FIXME
}
\examples{
get_subcollections()

}
